Abstract common OMQ/QUIC functionality
This moves most of the OMQ/QUIC request functionality into a new MQBase
class inherited by each, keeping just the OMQ- and QUIC-specific bits in
their respective classes.
