+++ oxenss/server/quic.cpp
@@ -1,74 +1,38 @@
 #include "quic.h"
-#include "omq.h"
+#include "oxenss/rpc/rate_limiter.h"
 
 #include <oxenss/rpc/request_handler.h>
 
-namespace oxenss::quic {
+namespace oxenss::server {
 
-Connection::Connection(
-        std::shared_ptr<oxen::quic::connection_interface>& c,
-        std::shared_ptr<oxen::quic::BTRequestStream>& s) :
-        conn{c}, control_stream{s} {}
+static auto logcat = log::Cat("ssquic");
 
-std::shared_ptr<Endpoint> Endpoint::make(
+QUIC::QUIC(
         rpc::RequestHandler& rh,
-        server::OMQ& q,
-        const Address& bind,
-        const crypto::ed25519_seckey& sk) {
-    std::shared_ptr<Endpoint> ep{new Endpoint{rh, q, bind, sk}};
-    return ep;
-}
-
-Endpoint::Endpoint(
-        rpc::RequestHandler& rh,
-        server::OMQ& q,
+        rpc::RateLimiter& rl,
         const Address& bind,
         const crypto::ed25519_seckey& sk) :
         local{bind},
         network{std::make_unique<oxen::quic::Network>()},
         tls_creds{oxen::quic::GNUTLSCreds::make_from_ed_seckey(sk.str())},
-        ep{create_endpoint()},
+        ep{network->endpoint(
+                local,
+                [this](oxen::quic::connection_interface& ci, uint64_t ec) {
+                    on_conn_closed(ci, ec);
+                })},
         request_handler{rh},
-        omq{q} {}
+        command_handler{[this](quic::message m) { handle_request(std::move(m)); }} {
+    request_handler_ = &rh;
+    rate_limiter_ = &rl;
+}
 
-std::shared_ptr<oxen::quic::Endpoint> Endpoint::create_endpoint() {
-    auto ep = network->endpoint(local, [this](oxen::quic::connection_interface& ci, uint64_t ec) {
-        on_conn_closed(ci, ec);
+void QUIC::startup_endpoint() {
+    ep->listen(tls_creds, [&](oxen::quic::connection_interface& c) {
+        c.queue_incoming_stream<oxen::quic::BTRequestStream>(command_handler);
     });
-    return ep;
 }
 
-void Endpoint::startup_endpoint() {
-    ep->listen(
-            tls_creds,
-            [&](oxen::quic::Connection& c,
-                oxen::quic::Endpoint& e,
-                std::optional<int64_t> id) -> std::shared_ptr<oxen::quic::Stream> {
-                if (id && id == 0) {
-                    auto s = std::make_shared<oxen::quic::BTRequestStream>(c, e);
-                    register_commands(s);
-                    return s;
-                }
-                return std::make_shared<oxen::quic::Stream>(c, e);
-            });
-}
-
-void Connection::send(std::string method, std::string body, quic_callback f) {
-    control_stream->command(std::move(method), std::move(body), std::move(f));
-}
-
-bool Endpoint::send(
-        oxen::quic::ConnectionID cid, std::string method, std::string body, quic_callback func) {
-
-    if (auto conn = get_conn(cid)) {
-        conn->control_stream->command(std::move(method), std::move(body), std::move(func));
-        return true;
-    }
-
-    return false;
-}
-
-void Endpoint::on_conn_closed(oxen::quic::connection_interface& conn_interface, uint64_t ec) {
+void QUIC::on_conn_closed(oxen::quic::connection_interface& conn_interface, uint64_t ec) {
     const auto& remote = conn_interface.local();
     const auto& scid = conn_interface.scid();
 
@@ -86,129 +50,49 @@ void Endpoint::on_conn_closed(oxen::quic::connection_interface& conn_interface, 
             remote);
 }
 
-void Endpoint::register_commands(std::shared_ptr<oxen::quic::BTRequestStream>& s) {
-    // register storage.{endpoint} commands
-    for (const auto& [n, cb] : rpc::RequestHandler::client_rpc_endpoints) {
-        std::string name{n};
-
-        s->register_command(name, [this, name](oxen::quic::message m) {
-            std::invoke(&Endpoint::handle_request, this, name, std::move(m), false);
-        });
-    }
-
-    s->register_command("monitor", [this](oxen::quic::message m) {
-        std::invoke(&Endpoint::handle_monitor_message, this, std::move(m));
-    });
+void QUIC::handle_monitor_message(oxen::quic::message m) {
+    handle_monitor(
+            m.body(),
+            [&m](std::string response) { m.respond(std::move(response)); },
+            m.stream()->conn_id());
 }
 
-void Endpoint::handle_monitor_message(oxen::quic::message m) {
-    if (m.timed_out) {
-        log::info(logcat, "Request (method:{}) timed out!");
+void QUIC::handle_request(oxen::quic::message m) {
+    auto name = m.endpoint();
+
+    if (handle_client_rpc(
+                name,
+                m.body(),
+                m.stream()->remote().host(),
+                [m = std::move(m)](http::response_code status, std::string_view body) {
+                    m.respond(body);
+                }))
         return;
-    }
 
-    std::string body{m.body()};
-    const auto& front = body.front();
-    std::string result;
-    std::vector<sub_info> subs;
+    if (name == "monitor")
+        return handle_monitor_message(std::move(m));
 
-    try {
-        if (front == 'd') {
-            oxenc::bt_dict_producer out;
-            handle_monitor_message_single(oxenc::bt_dict_consumer{body}, out, subs);
-            result = std::move(out).str();
-        } else {
-            oxenc::bt_list_producer out;
-            oxenc::bt_list_consumer l{body};
-            while (!l.is_finished())
-                handle_monitor_message_single(l.consume_dict_consumer(), out.append_dict(), subs);
-            result = std::move(out).str();
-        }
-    } catch (const std::exception& e) {
-        result = oxenc::bt_serialize(oxenc::bt_dict{
-                {"errcode", static_cast<int>(MonitorResponse::BAD_ARGS)},
-                {"error", "Invalid arguments: Failed to parse monitor.messages data value"}});
-        m.respond(std::move(result), true);
-        return;
-    }
-
-    if (not subs.empty())
-        omq.update_monitors(subs, get_conn(m.scid()));
-
-    m.respond(result);
+    throw quic::no_such_endpoint{};
 }
 
-void Endpoint::handle_request(std::string name, oxen::quic::message m, bool forwarded) {
-    if (m.timed_out) {
-        log::info(logcat, "Request (method:{}) timed out!");
-        return;
-    }
-
-    std::string err;
-    std::string params{m.body()};
-    auto bt_encoded = !params.empty() && params.front() == 'd';
-
-    // Check client rpc endpoints
-    if (auto itr = rpc::RequestHandler::client_rpc_endpoints.find(name);
-        itr != rpc::RequestHandler::client_rpc_endpoints.end()) {
-
-        // This endpoint shouldn't have been registered if it isn't in here:
-        assert(itr != rpc::RequestHandler::client_rpc_endpoints.end());
-
-        try {
-            itr->second.omq(
-                    request_handler,
-                    params,
-                    forwarded,
-                    [msg = std::move(m), bt_encoded](rpc::Response res) mutable {
-                        std::string body;
-
-                        if (auto* j = std::get_if<nlohmann::json>(&res.body)) {
-                            nlohmann::json resp;
-
-                            if (res.status != http::OK)
-                                resp = nlohmann::json::array({res.status.first, std::move(*j)});
-                            else
-                                resp = std::move(*j);
-
-                            if (bt_encoded)
-                                body = bt_serialize(json_to_bt(std::move(resp)));
-                            else
-                                body = resp.dump();
-                        } else
-                            body = view_body(res);
-
-                        std::string err;
-
-                        if (res.status == http::OK)
-                            err = fmt::format(
-                                    "OMQ RPC request successful, returning {}-byte {} response",
-                                    body.size(),
-                                    body.empty() ? "text"
-                                    : bt_encoded ? "bt-encoded"
-                                                 : "json");
-                        else
-                            err = fmt::format(
-                                    "OMQ RPC request failed, replying with [{}, {}]",
-                                    res.status.first,
-                                    body);
-
-                        log::debug(logcat, err);
-                        msg.respond(body, not(res.status == http::OK));
-                    });
-            return;
-        } catch (const rpc::parse_error& p) {
-            err = "Invalid request: {}"_format(p.what());
-            log::info(logcat, err);
-            err = serialize_error(BAD_REQUEST, std::move(err), bt_encoded);
-        } catch (const std::exception& e) {
-            err = "Client request raised an exception: {}"_format(e.what());
-            log::info(logcat, err);
-            err = serialize_error(INTERNAL_SERVER_ERROR, std::move(err), bt_encoded);
-        }
-
-        m.respond(err, true);
-    }
+nlohmann::json QUIC::wrap_response(
+        [[maybe_unused]] const http::response_code& status, nlohmann::json body) const {
+    // For QUIC requests we always wrap the result into a [CODE, BODY] list (even for successes).
+    // This is different from the OMQ because, in OMQ, messages are multi-part and so we can
+    // disambiguate success-with-body from failure-with-body by looking at the number of parts; here
+    // we can't, so we always make responses a 2-element list.
+    auto res = nlohmann::json::array();
+    res.push_back(status.first);
+    res.push_back(std::move(body));
+    return res;
 }
 
-}  // namespace oxenss::quic
+void QUIC::notify(std::vector<connection_id>& conns, std::string_view notification) {
+    for (const auto& c : conns)
+        if (auto* cid = std::get_if<oxen::quic::ConnectionID>(&c))  // FIXME: ConnRef or whatever
+            if (auto conn = ep->get_conn(*cid))
+                if (auto str = conn->get_stream<oxen::quic::BTRequestStream>(0))
+                    str->command("notify", notification);
+}
+
+}  // namespace oxenss::server
