+++ oxenss/server/quic.h
@@ -1,6 +1,8 @@
 #pragma once
 
+#include "oxenss/rpc/rate_limiter.h"
 #include "utils.h"
+#include "mqbase.h"
 
 #include <oxenss/crypto/keys.h>
 #include <oxenss/logging/oxen_logger.h>
@@ -12,129 +14,61 @@ class RequestHandler;
 }  // namespace oxenss::rpc
 
 namespace oxenss::server {
-class OMQ;
-}
 
-namespace oxenss::quic {
+namespace quic = oxen::quic;
 
-static auto logcat = log::Cat("quic");
+using quic_callback = std::function<void(quic::message)>;
+using Address = quic::Address;
 
-struct Endpoint;
-
-using quic_callback = std::function<void(oxen::quic::message)>;
-using Address = oxen::quic::Address;
-
-enum class message_type { REQUEST = 0, DATAGRAM = 1 };
-
-struct PendingMessage {
+struct PendingRequest {
     std::optional<std::string> name = std::nullopt;
     std::string body;
-    message_type type;
     quic_callback func = nullptr;
 
-    // Constructors for datagrams
-    PendingMessage(std::string _b) : body{std::move(_b)}, type{message_type::DATAGRAM} {}
-    PendingMessage(std::string_view _b) : body{_b}, type{message_type::DATAGRAM} {}
-
-    // Constructors for requests
-    PendingMessage(std::string _n, std::string _b, quic_callback _f) :
-            name{std::move(_n)},
-            body{std::move(_b)},
-            type{message_type::REQUEST},
-            func{std::move(_f)} {}
-    PendingMessage(std::string_view _n, std::string_view _b, quic_callback _f) :
-            name{_n}, body{_b}, type{message_type::REQUEST}, func{std::move(_f)} {}
+    // Constructor
+    PendingRequest(std::string name, std::string body, quic_callback func) :
+            name{std::move(name)}, body{std::move(body)}, func{std::move(func)} {}
+    PendingRequest(std::string_view name, std::string_view body, quic_callback func) :
+            name{name}, body{body}, func{std::move(func)} {}
 };
 
-using MessageQueue = std::deque<PendingMessage>;
+using RequestQueue = std::deque<PendingRequest>;
 
-struct Connection {
-    friend struct Endpoint;
-
-    Connection(
-            std::shared_ptr<oxen::quic::connection_interface>& c,
-            std::shared_ptr<oxen::quic::BTRequestStream>& s);
-
-    void send(std::string method, std::string body, quic_callback f = nullptr);
-
-  private:
-    std::shared_ptr<oxen::quic::connection_interface> conn;
-    std::shared_ptr<oxen::quic::BTRequestStream> control_stream;
-
+class QUIC : public MQBase {
   public:
-    //
-};
-
-struct Endpoint {
-
-    static std::shared_ptr<Endpoint> make(
-            rpc::RequestHandler& rh,
-            server::OMQ& q,
-            const Address& bind,
-            const crypto::ed25519_seckey& sk);
-
-    bool send(
-            oxen::quic::ConnectionID cid,
-            std::string method,
-            std::string body,
-            quic_callback func = nullptr);  // may not need this default values
-
-    std::shared_ptr<quic::Connection> get_conn(const oxen::quic::ConnectionID& cid) {
-        if (auto itr = conns.find(cid); itr != conns.end())
-            return itr->second;
-
-        return nullptr;
-    }
+    QUIC(rpc::RequestHandler& rh,
+         rpc::RateLimiter& rl,
+         const Address& bind,
+         const crypto::ed25519_seckey& sk);
 
     void startup_endpoint();
 
+    void notify(std::vector<connection_id>&, std::string_view notification) override;
+
   private:
-    Endpoint(
-            rpc::RequestHandler& rh,
-            server::OMQ& q,
-            const Address& bind,
-            const crypto::ed25519_seckey& sk);
 
     const Address local;
-    std::unique_ptr<oxen::quic::Network> network;
-    std::shared_ptr<oxen::quic::GNUTLSCreds> tls_creds;
-    std::shared_ptr<oxen::quic::Endpoint> ep;
+    std::unique_ptr<quic::Network> network;
+    std::shared_ptr<quic::GNUTLSCreds> tls_creds;
+    std::shared_ptr<quic::Endpoint> ep;
 
     rpc::RequestHandler& request_handler;
-    server::OMQ& omq;
+    std::function<void(quic::message m)> command_handler;
 
     // Holds all connections currently being managed by the quic endpoint
-    std::unordered_map<oxen::quic::ConnectionID, std::shared_ptr<quic::Connection>> conns;
+    std::unordered_map<quic::ConnectionID, std::shared_ptr<quic::Connection>> conns;
 
-    std::shared_ptr<oxen::quic::Endpoint> create_endpoint();
+    std::shared_ptr<quic::Endpoint> create_endpoint();
 
-    void on_conn_closed(oxen::quic::connection_interface& ci, uint64_t ec);
+    void on_conn_closed(quic::connection_interface& ci, uint64_t ec);
 
-    void register_commands(std::shared_ptr<oxen::quic::BTRequestStream>& s);
+    void handle_request(quic::message m);
 
-    void handle_request(std::string name, oxen::quic::message m, bool forwarded = false);
+    void handle_monitor_message(quic::message m);
 
-    void handle_monitor_message(oxen::quic::message m);
-
-  public:
-    template <typename... Opt>
-    bool establish_connection(const Address& addr, Opt&&... opts) {
-        try {
-            auto conn_interface = ep->connect(addr, tls_creds, std::forward<Opt>(opts)...);
-
-            // emplace immediately for connection open callback to find scid
-            auto [itr, b] = conns.emplace(conn_interface->scid(), nullptr);
-
-            auto control_stream =
-                    conn_interface->template get_new_stream<oxen::quic::BTRequestStream>();
-            itr->second = std::make_shared<quic::Connection>(conn_interface, control_stream);
-
-            return true;
-        } catch (...) {
-            log::error(logcat, "Error: failed to establish connection to {}", addr);
-            return false;
-        }
-    }
+    nlohmann::json wrap_response(
+            [[maybe_unused]] const http::response_code& status,
+            nlohmann::json response) const override;
 };
 
-}  // namespace oxenss::quic
+}  // namespace oxenss::server
