+++ oxenss/server/utils.cpp
@@ -75,165 +75,4 @@ nlohmann::json bt_to_json(oxenc::bt_list_consumer l) {
     return j;
 }
 
-void handle_monitor_message_single(
-        oxenc::bt_dict_consumer d, oxenc::bt_dict_producer& out, std::vector<sub_info>& subs) {
-
-    // Values we receive, in bt-dict order:
-    std::string_view ed_pk;                         // P (ed25519 pubkey only for session ids)
-    std::optional<signed_subaccount_token> subacc;  // S (token signature), T (token)
-    bool want_data = false;                         // d (given and true if data is desired)
-    using namespace_int = std::underlying_type_t<namespace_id>;
-    std::vector<namespace_id> namespaces;  // n (ordered list of numeric namespaces)
-    std::string pubkey;                    // p (network id + ed25519 pubkey; not a session id)
-    std::string_view signature;            // s (signature, by subacc.token.pubkey() or pubkey)
-    std::chrono::seconds timestamp;        // t
-
-    try {
-        // Ed25519 pubkey for a Session ID (required if `p` is not present)
-        if (d.skip_until("P")) {
-            ed_pk = d.consume_string_view();
-            if (ed_pk.size() != 32)
-                return monitor_error(
-                        out,
-                        MonitorResponse::BAD_PUBKEY,
-                        "Provided P= Session Ed25519 pubkey must be 32 bytes");
-        }
-
-        // Subaccount signature & token
-        if (d.skip_until("S")) {
-            subacc.emplace();
-            if (auto sig = d.consume_string_view(); sig.size() == subacc->signature.size())
-                std::memcpy(subacc->signature.data(), sig.data(), sig.size());
-            else
-                return monitor_error(
-                        out,
-                        MonitorResponse::BAD_PUBKEY,
-                        "Provided S subaccount signature must be {} bytes"_format(
-                                subacc->signature.size()));
-
-            if (auto token = d.require<std::string_view>("T");
-                token.size() == SUBACCOUNT_TOKEN_LENGTH)
-                std::memcpy(subacc->token.token.data(), token.data(), token.size());
-            else
-                return monitor_error(
-                        out,
-                        MonitorResponse::BAD_PUBKEY,
-                        "Provided T subaccount token must be {} bytes"_format(
-                                SUBACCOUNT_TOKEN_LENGTH));
-        }
-
-        // Flag to send full message data as part of the pushed notifications (optional)
-        if (d.skip_until("d"))
-            want_data = d.consume_integer<bool>();
-
-        // List of namespaces to monitor (required)
-        auto ns = d.require<oxenc::bt_list_consumer>("n");
-        namespaces.push_back(static_cast<namespace_id>(ns.consume_integer<namespace_int>()));
-        while (!ns.is_finished()) {
-            auto nsi = static_cast<namespace_id>(ns.consume_integer<namespace_int>());
-            if (nsi > namespaces.back())
-                namespaces.push_back(nsi);
-            else
-                return monitor_error(
-                        out,
-                        MonitorResponse::BAD_NS,
-                        "Invalid n= namespace list: namespaces must be ascending");
-        }
-
-        // Account ID (i.e. network prefix followed by an Ed25519 pubkey).  *Not* for Session
-        // IDs, which must use P instead (because Session IDs are X25519 pubkeys).
-        if (d.skip_until("p")) {
-            if (!ed_pk.empty())
-                throw std::runtime_error{"Cannot provide both p= and P= pubkey values"};
-            pubkey = d.consume_string();
-            if (pubkey.size() != 33)
-                monitor_error(
-                        out, MonitorResponse::BAD_PUBKEY, "Provided p= pubkey must be 33 bytes");
-        } else if (ed_pk.empty()) {
-            throw std::runtime_error{"Either p= or P= must be given"};
-        }
-
-        signature = d.require<std::string_view>("s");
-        if (signature.size() != 64)
-            return monitor_error(
-                    out, MonitorResponse::BAD_SIG, "Provided s= signature must be 64 bytes");
-
-        if (!d.skip_until("t"))
-            throw std::runtime_error{"required signature timestamp is missing"};
-        timestamp = std::chrono::seconds{d.consume_integer<int64_t>()};
-
-    } catch (const std::exception& ex) {
-        return monitor_error(out, MonitorResponse::BAD_ARGS, "Invalid arguments: "s + ex.what());
-    }
-
-    // Make sure the sig timestamp isn't too old or too new
-    auto now = std::chrono::system_clock::now();
-    auto ts = std::chrono::system_clock::time_point(timestamp);
-    if (bool too_old = ts < now - 14 * 24h; too_old || ts > now + 24h) {
-        return monitor_error(
-                out,
-                MonitorResponse::BAD_TS,
-                "Invalid t= signature timestamp: timestamp is "s +
-                        (too_old ? "too old" : "in the future"));
-    }
-
-    // If given an Ed25519 pubkey for a Session ID, derive the Session ID
-    if (!ed_pk.empty()) {
-        pubkey.resize(33);
-        pubkey[0] = 0x05;
-        if (auto rc = crypto_sign_ed25519_pk_to_curve25519(
-                    reinterpret_cast<unsigned char*>(pubkey.data() + 1),
-                    reinterpret_cast<const unsigned char*>(ed_pk.data()));
-            rc != 0)
-            return monitor_error(out, MonitorResponse::BAD_PUBKEY, "Invalid P= ed25519 public key");
-    } else {
-        // No Session Ed25519, so assume the pubkey (without prefix byte) is Ed25519
-        ed_pk = pubkey;
-        ed_pk.remove_prefix(1);
-    }
-
-    std::string_view verify_key = ed_pk;
-    if (subacc) {
-        try {
-            subacc->verify(
-                    pubkey[0],
-                    reinterpret_cast<const unsigned char*>(ed_pk.data()),
-                    subaccount_access::Read);
-        } catch (const std::exception& ex) {
-            auto m = "Subaccount verification failed: {}"_format(ex.what());
-            log::warning(logcat, "{}", m);
-            return monitor_error(out, MonitorResponse::BAD_SIG, m);
-        }
-        auto sub_pk = subacc->token.pubkey();
-        verify_key = std::string_view{reinterpret_cast<const char*>(sub_pk.data()), sub_pk.size()};
-    }
-    assert(verify_key.size() == 32);
-
-    auto pubkey_hex = oxenc::to_hex(pubkey);
-
-    auto sig_msg = fmt::format(
-            "MONITOR{:s}{:d}{:d}{}",
-            pubkey_hex,
-            timestamp.count(),
-            want_data,
-            fmt::join(namespaces, ","));
-
-    if (0 != crypto_sign_verify_detached(
-                     reinterpret_cast<const unsigned char*>(signature.data()),
-                     reinterpret_cast<const unsigned char*>(sig_msg.data()),
-                     sig_msg.size(),
-                     reinterpret_cast<const unsigned char*>(verify_key.data()))) {
-        log::debug(logcat, "monitor.messages signature verification failed");
-        return monitor_error(out, MonitorResponse::BAD_SIG, "Signature verification failed");
-    }
-
-    subs.emplace_back(std::move(pubkey), std::move(pubkey_hex), std::move(namespaces), want_data);
-    out.append("success", 1);
-}
-
-void handle_monitor_message_single(
-        oxenc::bt_dict_consumer d, oxenc::bt_dict_producer&& out, std::vector<sub_info>& subs) {
-    handle_monitor_message_single(d, out, subs);
-}
-
 }  // namespace oxenss
