+++ unit_test/CMakeLists.txt
@@ -15,7 +15,7 @@ add_executable(Test
 
 target_link_libraries(Test
     PRIVATE
-    common storage utils crypto snode rpc
+    common storage utils crypto snode rpc server
     Catch2::Catch2)
 
 target_include_directories(Test PRIVATE ..)
